# api/openapi.json
{
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action": {
            "enum": [
              "save",
              "settle"
            ],
            "title": "Action",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
...
      "BroadcastRequest": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "BroadcastRequest",
        "type": "object"
      },
      "CommandRequest": {
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          }
        },
        "required": [
          "command"
        ],
        "title": "CommandRequest",
        "type": "object"
      },
      "ConfigUpdateRequest": {
        "properties": {
          "apply": {
            "default": false,
            "title": "Apply",
            "type": "boolean"
          },
          "confirm_low_max_players": {
            "default": false,
            "title": "Confirm Low Max Players",
            "type": "boolean"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "ConfigUpdateRequest",
        "type": "object"
      },
      "ConfigUpdateResponse": {
        "properties": {
          "applied": {
            "items": {
              "type": "string"
            },
            "title": "Applied",
            "type": "array"
          },
          "changed": {
            "items": {
              "type": "string"
            },
            "title": "Changed",
            "type": "array"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "persisted": {
            "items": {
              "type": "string"
            },
            "title": "Persisted",
            "type": "array"
          },
          "requires_restart": {
            "items": {
              "type": "string"
            },
            "title": "Requires Restart",
            "type": "array"
          }
        },
        "required": [
          "persisted",
          "changed",
          "applied",
          "requires_restart"
        ],
        "title": "ConfigUpdateResponse",
        "type": "object"
      },
      "ConfigView": {
        "description": "serverconfig.txt 里可编辑的键（world 已归一化成文件名）。",
        "properties": {
          "editable_keys": {
            "items": {
              "type": "string"
            },
            "title": "Editable Keys",
            "type": "array"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "restart_keys": {
            "items": {
              "type": "string"
            },
            "title": "Restart Keys",
            "type": "array"
          },
          "runtime_keys": {
            "items": {
              "type": "string"
            },
            "title": "Runtime Keys",
            "type": "array"
          },
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values",
          "editable_keys",
          "runtime_keys",
          "restart_keys",
          "path"
        ],
        "title": "ConfigView",
        "type": "object"
      },
      "ConsoleLine": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "offset",
          "kind",
          "text"
        ],
        "title": "ConsoleLine",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "TimeRequest": {
        "properties": {
          "phase": {
            "enum": [
              "dawn",
              "noon",
              "dusk",
              "midnight"
            ],
            "title": "Phase",
            "type": "string"
          }
        },
        "required": [
          "phase"
        ],
        "title": "TimeRequest",
        "type": "object"
      },
      "TimeResponse": {
        "properties": {
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "required": [
          "phase",
          "time"
        ],
        "title": "TimeResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__schemas__v1__CommandResponse": {
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "output": {
            "title": "Output",
            "type": "string"
          }
        },
        "required": [
          "command",
          "output"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
      "app__schemas__v1__ConsoleResponse": {
        "properties": {
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ConsoleLine"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "lines",
          "cursor"
        ],
        "title": "ConsoleResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Terraria Server API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/broadcast": {
      "post": {
        "operationId": "broadcast_api_v1_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Broadcast",
        "tags": [
          "v1:players"
        ]
      }
    },
    "/api/v1/config": {
      "get": {
        "operationId": "get_config_api_v1_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "v1:server"
        ]
      },
      "put": {
        "operationId": "update_config_api_v1_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "v1:server"
        ]
      }
    },
    "/api/v1/console": {
      "get": {
        "description": "`since` 缺省时返回最后 tail 行；给了 since 就返回该字节偏移之后的新行。",
        "operationId": "console_api_v1_console_get",
        "parameters": [
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Tail",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__v1__ConsoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Console",
        "tags": [
          "v1:console"
        ]
      }
    },
...
    "/api/v1/console/commands": {
      "post": {
        "operationId": "run_command_api_v1_console_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__v1__CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Command",
        "tags": [
          "v1:console"
        ]
      }
    },
...
    "/api/v1/players/{name}/ban": {
      "delete": {
        "operationId": "unban_player_api_v1_players__name__ban_delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unban Player",
        "tags": [
          "v1:players"
        ]
      },
      "post": {
        "operationId": "ban_player_api_v1_players__name__ban_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ban Player",
        "tags": [
          "v1:players"
        ]
      }
    },
    "/api/v1/players/{name}/kick": {
      "post": {
        "operationId": "kick_player_api_v1_players__name__kick_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kick Player",
        "tags": [
          "v1:players"
        ]
      }
    },
...
    "/api/v1/server/actions": {
      "post": {
        "operationId": "server_action_api_v1_server_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server Action",
        "tags": [
          "v1:server"
        ]
      }
    },
...
    "/api/v1/server/time": {
      "post": {
        "operationId": "set_time_api_v1_server_time_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Time",
        "tags": [
          "v1:server"
        ]
      }
    },
...
  }
}